# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SQLoptimize API",
    "description": "SQLoptimize API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/databases/active": {
      "get": {
        "tags": [
          "databases"
        ],
        "summary": "获取用户当前活跃的数据库连接",
        "description": "获取用户当前活跃的数据库连接",
        "operationId": "get_active_database_connection_api_databases_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveDatabaseConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "databases"
        ],
        "summary": "设置用户当前活跃的数据库连接",
        "description": "设置用户当前活跃的数据库连接",
        "operationId": "set_active_database_connection_api_databases_active_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDatabaseConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetActiveDatabaseConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/databases/": {
      "get": {
        "tags": [
          "databases"
        ],
        "summary": "获取用户数据库连接列表",
        "description": "获取当前用户的数据库连接列表",
        "operationId": "get_user_databases_api_databases__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "database_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Database Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "databases"
        ],
        "summary": "创建数据库连接",
        "description": "创建新的数据库连接",
        "operationId": "create_database_connection_api_databases__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/databases/{connection_id}": {
      "get": {
        "tags": [
          "databases"
        ],
        "summary": "根据ID获取数据库连接",
        "description": "获取指定的数据库连接信息",
        "operationId": "get_database_connection_api_databases__connection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "databases"
        ],
        "summary": "更新数据库连接",
        "description": "更新数据库连接信息",
        "operationId": "update_database_connection_api_databases__connection_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "databases"
        ],
        "summary": "删除数据库连接",
        "description": "删除数据库连接",
        "operationId": "delete_database_connection_api_databases__connection_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/databases/{connection_id}/test": {
      "post": {
        "tags": [
          "databases"
        ],
        "summary": "测试数据库连接",
        "description": "测试数据库连接",
        "operationId": "test_database_connection_api_databases__connection_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionTestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveDatabaseConnectionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "数据库连接ID"
          },
          "database_name": {
            "type": "string",
            "title": "Database Name",
            "description": "用户自定义的数据库名称"
          },
          "database_uri": {
            "type": "string",
            "title": "Database Uri",
            "description": "数据库连接URI"
          },
          "database_type": {
            "type": "string",
            "title": "Database Type",
            "description": "数据库类型"
          },
          "database_description": {
            "type": "string",
            "title": "Database Description",
            "description": "数据库描述"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "database_name",
          "database_uri",
          "database_type",
          "database_description",
          "created_at",
          "updated_at"
        ],
        "title": "ActiveDatabaseConnectionResponse",
        "description": "活跃数据库连接响应"
      },
...
      "DatabaseConnectionCreate": {
        "properties": {
          "database_name": {
            "type": "string",
            "title": "Database Name",
            "description": "用户自定义的数据库名称"
          },
          "database_uri": {
            "type": "string",
            "title": "Database Uri",
            "description": "数据库连接URI"
          },
          "database_type": {
            "type": "string",
            "title": "Database Type",
            "description": "数据库类型",
            "default": "opentenbase"
          },
          "database_description": {
            "type": "string",
            "title": "Database Description",
            "description": "数据库描述",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "database_name",
          "database_uri"
        ],
        "title": "DatabaseConnectionCreate",
        "description": "创建数据库连接请求"
      },
      "DatabaseConnectionDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "删除结果消息"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "DatabaseConnectionDeleteResponse",
        "description": "数据库连接删除响应"
      },
      "DatabaseConnectionListResponse": {
        "properties": {
          "databases": {
            "items": {
              "$ref": "#/components/schemas/DatabaseConnectionResponse"
            },
            "type": "array",
            "title": "Databases",
            "description": "数据库连接列表"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总数"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "跳过数量"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "限制数量"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "是否还有更多数据"
          },
          "active_connection_id": {
            "type": "integer",
            "title": "Active Connection Id",
            "description": "当前用户活跃的数据库连接ID，0表示无活跃连接",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "databases",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "DatabaseConnectionListResponse",
        "description": "数据库连接列表响应"
      },
      "DatabaseConnectionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "数据库连接ID"
          },
          "database_name": {
            "type": "string",
            "title": "Database Name",
            "description": "用户自定义的数据库名称"
          },
          "database_uri": {
            "type": "string",
            "title": "Database Uri",
            "description": "数据库连接URI"
          },
          "database_type": {
            "type": "string",
            "title": "Database Type",
            "description": "数据库类型"
          },
          "database_description": {
            "type": "string",
            "title": "Database Description",
            "description": "数据库描述"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "database_name",
          "database_uri",
          "database_type",
          "database_description",
          "created_at",
          "updated_at"
        ],
        "title": "DatabaseConnectionResponse",
        "description": "数据库连接响应"
      },
      "DatabaseConnectionTestResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "测试结果消息"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "测试状态：success/failed"
          },
          "response_time": {
            "type": "number",
            "title": "Response Time",
            "description": "响应时间"
          }
        },
        "type": "object",
        "required": [
          "message",
          "status",
          "response_time"
        ],
        "title": "DatabaseConnectionTestResponse",
        "description": "数据库连接测试响应"
      },
      "DatabaseConnectionUpdate": {
        "properties": {
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name",
            "description": "用户自定义的数据库名称"
          },
          "database_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Uri",
            "description": "数据库连接URI"
          },
          "database_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Type",
            "description": "数据库类型"
          },
          "database_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Description",
            "description": "数据库描述"
          }
        },
        "type": "object",
        "title": "DatabaseConnectionUpdate",
        "description": "更新数据库连接请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetActiveDatabaseConnectionRequest": {
        "properties": {
          "connection_id": {
            "type": "integer",
            "title": "Connection Id",
            "description": "数据库连接ID"
          }
        },
        "type": "object",
        "required": [
          "connection_id"
        ],
        "title": "SetActiveDatabaseConnectionRequest",
        "description": "设置活跃数据库连接请求"
      },
      "SetActiveDatabaseConnectionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "设置结果消息"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SetActiveDatabaseConnectionResponse",
        "description": "设置活跃数据库连接响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}