# backend/src/api/router/database_router.py
from fastapi import APIRouter, HTTPException, Depends, status
from typing import Optional
from src.schemas.repository.database import CreateDatabaseConnectionReq
from src.api.repository import DatabaseConnectionRepository
from src.api.utils import get_current_user
from src.db.registry import DatabaseRegistry
from sqlalchemy import text
import logging
import time
from src.schemas.params.database import (
    ActiveDatabaseConnectionResponse, 
    SetActiveDatabaseConnectionRequest, 
    SetActiveDatabaseConnectionResponse, 
    DatabaseConnectionListResponse, 
    DatabaseConnectionResponse, 
    DatabaseConnectionUpdate, 
    DatabaseConnectionCreate,
    DatabaseConnectionDeleteResponse,
    DatabaseConnectionTestResponse
)
logger = logging.getLogger(__name__)

database_router = APIRouter(
    prefix="/api/databases",
    tags=["databases"],
    responses={404: {"description": "Not found"}},
)

@database_router.get("/active", response_model=ActiveDatabaseConnectionResponse, summary="获取用户当前活跃的数据库连接")
async def get_active_database_connection(
    current_user: dict = Depends(get_current_user)
):
    """获取用户当前活跃的数据库连接"""
    try:
        db_repo = DatabaseConnectionRepository()
        active_connection = db_repo.get_active_by_user_id(current_user["id"])
        
        if not active_connection:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail="用户没有任何数据库连接"
            )
        
        # 转换为响应模型
        response = ActiveDatabaseConnectionResponse(
            id=active_connection.id,
            database_name=active_connection.database_name,
            database_uri=active_connection.database_uri,
            database_type=active_connection.database_type,
            database_description=active_connection.database_description,
            created_at=active_connection.created_at.isoformat() if active_connection.created_at else "",
            updated_at=active_connection.updated_at.isoformat() if active_connection.updated_at else ""
        )
        
        logger.info(f"获取用户活跃数据库连接成功，用户ID: {current_user['id']}, 连接ID: {active_connection.id}")
        return response
            
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"获取用户活跃数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取活跃数据库连接失败，请稍后重试"
        )

@database_router.post("/active", response_model=SetActiveDatabaseConnectionResponse, summary="设置用户当前活跃的数据库连接")
async def set_active_database_connection(
    request: SetActiveDatabaseConnectionRequest,
    current_user: dict = Depends(get_current_user)
):
    """设置用户当前活跃的数据库连接"""
    try:
        db_repo = DatabaseConnectionRepository()
        
        # 设置活跃连接
        success = db_repo.set_active_by_user_id(current_user["id"], request.connection_id)
        
        if not success:
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="数据库连接不存在或不属于当前用户"
            )
        
        logger.info(f"设置用户活跃数据库连接成功，用户ID: {current_user['id']}, 连接ID: {request.connection_id}")
        return SetActiveDatabaseConnectionResponse(message="活跃数据库连接设置成功")
            
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"设置用户活跃数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="设置活跃数据库连接失败，请稍后重试"
        )

@database_router.get("/", response_model=DatabaseConnectionListResponse, summary="获取用户数据库连接列表")
async def get_user_databases(
    skip: int = 0,
    limit: int = 100,
    database_type: Optional[str] = None,
    current_user: dict = Depends(get_current_user)
):
    """获取当前用户的数据库连接列表"""
    try:
        db_repo = DatabaseConnectionRepository()
        if database_type:
            connections = db_repo.get_by_user_and_type(current_user["id"], database_type)
        else:
            connections = db_repo.get_by_user_id(current_user["id"], skip, limit)
        
        # 转换为响应模型
        database_responses = []
        for conn in connections:
            database_responses.append(DatabaseConnectionResponse(
                id=conn.id,
                database_name=conn.database_name,
                database_uri=conn.database_uri,
                database_type=conn.database_type,
                database_description=conn.database_description,
                created_at=conn.created_at.isoformat() if conn.created_at else "",
                updated_at=conn.updated_at.isoformat() if conn.updated_at else ""
            ))
        
        # 获取总数（用于分页）
        total = db_repo.count_by_user_id(current_user["id"])
        
        # 计算是否还有更多数据
        has_more = (skip + len(database_responses)) < total
        
        # 获取当前用户活跃的数据库连接ID
        active_connection = db_repo.get_active_by_user_id(current_user["id"], auto_set_first=True)
        active_connection_id = active_connection.id if active_connection else 0
        
        logger.info(f"获取用户数据库连接列表成功，用户ID: {current_user['id']}, 数量: {len(database_responses)}")
        
        return DatabaseConnectionListResponse(
            databases=database_responses,
            total=total,
            skip=skip,
            limit=limit,
            has_more=has_more,
            active_connection_id=active_connection_id
        )
            
    except Exception as e:
        logger.error(f"获取用户数据库连接列表失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取数据库连接列表失败，请稍后重试"
        )

@database_router.post("/", response_model=DatabaseConnectionResponse, summary="创建数据库连接")
async def create_database_connection(
    request: DatabaseConnectionCreate,
    current_user: dict = Depends(get_current_user)
):
    """创建新的数据库连接"""
    try:
        db_repo = DatabaseConnectionRepository()
        
        # 检查数据库名称是否已存在
        if db_repo.exists_by_name(request.database_name):
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="数据库名称已存在"
            )
        
        # 创建数据库连接
        connection_data = CreateDatabaseConnectionReq(
            database_name=request.database_name,
            database_uri=request.database_uri,
            database_type=request.database_type,
            database_description=request.database_description,
            user_id=current_user["id"]
        )
        
        connection = db_repo.create(connection_data)
        
        # 转换为响应模型
        response = DatabaseConnectionResponse(
            id=connection.id,
            database_name=connection.database_name,
            database_uri=connection.database_uri,
            database_type=connection.database_type,
            database_description=connection.database_description,
            created_at=connection.created_at.isoformat() if connection.created_at else "",
            updated_at=connection.updated_at.isoformat() if connection.updated_at else ""
        )
        
        logger.info(f"创建数据库连接成功，用户ID: {current_user['id']}, 数据库名称: {connection.database_name}")
        return response
            
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"创建数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="创建数据库连接失败，请稍后重试"
        )

@database_router.get("/{connection_id}", response_model=DatabaseConnectionResponse, summary="根据ID获取数据库连接")
async def get_database_connection(
    connection_id: int,
    current_user: dict = Depends(get_current_user)
):
    """获取指定的数据库连接信息"""
    try:
        db_repo = DatabaseConnectionRepository()
        connection = db_repo.get_by_id(connection_id)
        
        if not connection:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail="数据库连接不存在"
            )
        
        # 检查是否属于当前用户
        if connection.user_id != current_user["id"]:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="无权访问此数据库连接"
            )
        
        # 转换为响应模型
        response = DatabaseConnectionResponse(
            id=connection.id,
            database_name=connection.database_name,
            database_uri=connection.database_uri,
            database_type=connection.database_type,
            database_description=connection.database_description,
            created_at=connection.created_at.isoformat() if connection.created_at else "",
            updated_at=connection.updated_at.isoformat() if connection.updated_at else ""
        )
        
        logger.info(f"获取数据库连接成功，用户ID: {current_user['id']}, 连接ID: {connection_id}")
        return response
            
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"获取数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="获取数据库连接失败，请稍后重试"
        )

@database_router.put("/{connection_id}", response_model=DatabaseConnectionResponse, summary="更新数据库连接")
async def update_database_connection(
    connection_id: int,
    request: DatabaseConnectionUpdate,
    current_user: dict = Depends(get_current_user)
):
    """更新数据库连接信息"""
    try:
        db_repo = DatabaseConnectionRepository()
        
        # 检查数据库连接是否存在
        connection = db_repo.get_by_id(connection_id)
        if not connection:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail="数据库连接不存在"
            )
        
        # 检查是否属于当前用户
        if connection.user_id != current_user["id"]:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="无权修改此数据库连接"
            )
        
        # 如果更新数据库名称，检查是否与其他连接冲突
        if request.database_name and request.database_name != connection.database_name:
            if db_repo.exists_by_name(request.database_name):
                raise HTTPException(
                    status_code=status.HTTP_400_BAD_REQUEST,
                    detail="数据库名称已存在"
                )
        
        # 准备更新数据（只包含非None的字段）
        update_data = {}
        for field, value in request.model_dump(exclude_unset=True).items():
            if value is not None:
                update_data[field] = value
        
        if not update_data:
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="没有提供要更新的字段"
            )
        
        # 更新数据库连接
        updated_connection = db_repo.update(connection_id, update_data)
        
        # 转换为响应模型
        response = DatabaseConnectionResponse(
            id=updated_connection.id,
            database_name=updated_connection.database_name,
            database_uri=updated_connection.database_uri,
            database_type=updated_connection.database_type,
            database_description=updated_connection.database_description,
            created_at=updated_connection.created_at.isoformat() if updated_connection.created_at else "",
            updated_at=updated_connection.updated_at.isoformat() if updated_connection.updated_at else ""
        )
        
        logger.info(f"更新数据库连接成功，用户ID: {current_user['id']}, 连接ID: {connection_id}")
        return response
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"更新数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="更新数据库连接失败，请稍后重试"
        )

@database_router.delete("/{connection_id}", response_model=DatabaseConnectionDeleteResponse, summary="删除数据库连接")
async def delete_database_connection(
    connection_id: int,
    current_user: dict = Depends(get_current_user)
):
    """删除数据库连接"""
    try:
        db_repo = DatabaseConnectionRepository()
        
        # 检查数据库连接是否存在
        connection = db_repo.get_by_id(connection_id)
        if not connection:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail="数据库连接不存在"
            )
        
        # 检查是否属于当前用户
        if connection.user_id != current_user["id"]:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="无权删除此数据库连接"
            )
        
        # 删除数据库连接
        success = db_repo.delete(connection_id)
        if not success:
            raise HTTPException(
                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                detail="删除数据库连接失败"
            )
        
        logger.info(f"删除数据库连接成功，用户ID: {current_user['id']}, 连接ID: {connection_id}")
        return DatabaseConnectionDeleteResponse(message="数据库连接删除成功")
            
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"删除数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="删除数据库连接失败，请稍后重试"
        )

@database_router.post("/{connection_id}/test", response_model=DatabaseConnectionTestResponse, summary="测试数据库连接")
async def test_database_connection(
    connection_id: int,
    current_user: dict = Depends(get_current_user)
):
    """测试数据库连接"""
    try:
        db_repo = DatabaseConnectionRepository()
        db_connection = db_repo.get_by_id(connection_id)
        if not db_connection:
            raise HTTPException(
                status_code=status.HTTP_404_NOT_FOUND,
                detail="数据库连接不存在"
            )
        db_registry = DatabaseRegistry(db_connection.database_uri)
        # 检查数据库是否属于当前用户
        if db_connection.user_id != current_user["id"]:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail="无权测试此数据库连接"
            )

        with db_registry.session() as db:
            start_time = time.time()
            result = db.execute(text("SELECT 1")) # 测试连接
            logger.info(f"查询结果: {result.fetchone()}")
            response_time = time.time() - start_time
            logger.info(f"数据库连接测试成功，用户ID: {current_user['id']}, 连接ID: {connection_id}, 响应时间: {response_time}")
        return DatabaseConnectionTestResponse(
            message="数据库连接测试成功", 
            status="success", 
            response_time=response_time
        )
    except Exception as e:
        logger.error(f"测试数据库连接失败: {e}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="测试数据库连接失败，请稍后重试"
        )